generalized LdbRdfPropertySelectorController and used it also in in LocationFromSubject controller (#37), therefore renamed property for abstract to comment, because of its rdf name
