+++ app/controllers/MetainfoFromSubject.scala
@@ -7,16 +7,7 @@ import play.api.libs.json.Json
 import play.api.libs.json.Json._
 import play.api.mvc.{Result, Action}
 
-object MetainfoFromSubject extends LdbController {
+object MetainfoFromSubject extends LdbRdfPropertySelectorController {
 
-  val rdfsComment: Property = ResourceFactory.createProperty("http://www.w3.org/2000/01/rdf-schema#", "comment")
-
-  override def process(uri: String, endpoint: SparqlEndpoint, m: Model): Result = {
-    val givenResource: Resource = ResourceFactory.createResource(uri)
-    val returnable: String = Option(m.getProperty(givenResource, rdfsComment)).map(_.getLiteral.getLexicalForm).getOrElse({
-      logger.warn(s"No abstract found for uri ${uri}")
-        ""
-    })
-    Ok(toJson(Map("abstract" -> returnable)))
-  }
+  override def selectablePropertyUris: List[String] = "http://www.w3.org/2000/01/rdf-schema#comment" :: Nil
 }
